packages/stargate-mcp-ui-server/pyproject.toml
[project]
name = "stargate-mcp-ui-server"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "mcp[server]>=1.5.0",
    "mcp-ui-server>=1.0.0",
    "uvicorn[standard]>=0.29.0",
    "python-dotenv>=1.0.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/stargate_mcp_ui_server"]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[dependency-groups]
dev = [
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
]
